+++ smartsim/_core/schemas/dragonRequests.py
@@ -47,6 +47,8 @@ class DragonRunPolicy(BaseModel):
     """List of CPU indices to which the job should be pinned"""
     gpu_affinity: t.List[NonNegativeInt] = Field(default_factory=list)
     """List of GPU indices to which the job should be pinned"""
+    hostlist: t.List[str] = Field(default_factory=list)
+    """List of host names on which the job should be executed"""
 
     @staticmethod
     def from_run_args(
@@ -64,14 +66,20 @@ class DragonRunPolicy(BaseModel):
         if cpu_arg_value := run_args.get("cpu-affinity", None):
             cpu_args = str(cpu_arg_value)
 
-        # run args converted to a string must be split back into a list[int]
+        # list[int] converted to comma-separated str must split into a list[int]
         gpu_affinity = [int(x.strip()) for x in gpu_args.split(",") if x]
         cpu_affinity = [int(x.strip()) for x in cpu_args.split(",") if x]
+
+        # # list[str] converted to comma-separated str must split into a list[str]
+        # hosts: t.List[str] = []
+        # if hostlist_value := str(run_args.get("host-list", "")):
+        #     hosts = [x for x in hostlist_value.split(",") if x]
 
         try:
             return DragonRunPolicy(
                 cpu_affinity=cpu_affinity,
                 gpu_affinity=gpu_affinity,
+                # hostlist=hosts,
             )
         except ValidationError as ex:
             raise SmartSimError("Unable to build DragonRunPolicy") from ex